requirements.txt
# Core ML/DL
tensorflow>=2.13.0 
numpy>=1.24.3
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.3
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.1
seaborn>=0.11.2

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Progress bars
tqdm>=4.64.0

# Development tools
pytest>=7.4.0
black>=22.3.0
flake8>=6.0.0

# Optional - for notebooks
ipywidgets>=8.0.0
jupyter>=1.0.0 